backend/requirements.txt
# TitanNVR Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
asyncpg==0.29.0
alembic==1.12.1
pyyaml==6.0.1
docker==7.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Email notifications
aiosmtplib==3.0.1
email-validator==2.1.0
